Enable development on localhost
Relax the https requirement, because we usually
develop with only http on localhost. (TLS termination
in a business application is quite complex to setup)

This change will let us create a separate ID-porten client,
with separate secrets, which can be used for
developing a NAV application on localhost, while
integrating with ID-porten for testing.