fix(crypto): only set `kid` header instead of both `x5c` and `kid` for JWT grants
The `kid` header must be equal to the `kid` for the pre-registered
public JWK at DigDir.

The `x5c` header in the JWT grants is unnecessary if the pre-registered
JWK already includes the `x5c` parameter. DigDir already takes care of
validating the certificate chain for said key at runtime.